refactor: migrate from 32-bit to 64-bit fixed-width ISA
Replace the 32-bit instruction format (with WIDE prefix for large operands)
with a native 64-bit format: [op:8][_:8][A:16][B:16][C:16] for ABC,
[op:8][_:8][A:16][Bx:32] for ABx. This eliminates the WIDE prefix entirely,
simplifying the dispatch loop, compiler emitter, disassembler, verifier,
and serializer.

Key changes:
- Uint32List → Uint64List for bytecode throughout
- Op.wide deleted; all encodeWide*/decodeWide* functions removed
- sBx bias: 0x7FFF → 0x7FFFFFFF (32-bit); sAx bias: 0x7FFFFF → 0x7FFFFFFFFFFF
- Jump placeholders: 1-word (was 3-word with WIDE prefix)
- .darb format version: 2 → 3
- Fix latent RETHROW bug: -1 stack trace sentinel now emits 0 instead of
  wrapping to 0xFFFF in the wider B field

60 files changed, ~850 net lines removed.

+++ lib/src/bytecode/module.dart
@@ -142,8 +142,8 @@ class DarticFuncProto {
   /// and global function table indexing.
   final int funcId;
 
-  /// 32-bit fixed-width instruction sequence.
-  final Uint32List bytecode;
+  /// 64-bit fixed-width instruction sequence.
+  final Uint64List bytecode;
 
   /// Value stack frame space requirement (params + locals + temporaries).
   final int valueRegCount;
